Continue organizing, renaming and relinking
