+++ src/Components/FishCard/index.js
@@ -9,7 +9,7 @@ class FishCard extends React.Component {
     return (
       <div className="critter">
         <img
-          src={require(`../data/images/Fish/${this.props.image}.png`)}
+          src={require(`../../data/images/Fish/${this.props.image}.png`)}
           alt="fish"
         ></img>
         <p className="critter-name">{this.props.name}</p>
