+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./Pages/App";
+import App from "../src/Pages/App/";
 
 ReactDOM.render(<App />, document.querySelector("#root"));
